+++ packages/polythene-css-core/src/index.js
@@ -1,0 +1,28 @@
+// @ts-check
+
+import reset from "./reset";
+import { styler } from "polythene-core-css";
+
+const fns = [reset];
+const selector = "";
+const vars = {};
+
+const addStyle = styler.createAddStyle(selector, fns, vars);
+
+const getStyle = styler.createGetStyle(selector, fns, vars);
+
+const addGeneralStyleToHead = () =>
+  styler.addStyle({
+    identifier: "pe-core",
+    selectors: [selector],
+    fns,
+    vars
+  });
+
+export {
+  addGeneralStyleToHead,
+  addStyle,
+  getStyle,
+  reset,
+  vars,
+};